TypeQL 3.0 (#342)
## Usage and product changes

User-defined functions and structs:
```typeql
fun mean_salary($c: company) -> double? :
    match 
        (company: $c, employee: $_) isa employment, has salary $s;
    return mean($s); 
```
```typeql
struct dated_coordinate:
    longitude value double,
    latitude value double,
    date value datetime;
```

Query pipelines:
```
with fun costliest_printer($employee: employee) -> printer? :
  match 
    ($printer, $employee) isa print_permission;
    $printer has cost_per_page $cost;
  sort $cost desc;
  return first($printer);
match
  $printer isa printer, has office_number $n, has newly_installed true;
  $employee isa employee, has office_number $n;
put ($employee, $printer) isa print_permission;
match 
  $high_cost_printer = costliest_printer($employee), has printer_name $name;
  not { $printer is $high_cost_printer; };
  $employee has contact $address;
insert 
  $notice isa queued_email, has recipient $address, 
  has content "Do you still need the printer " + $name + "?";
```

New undefine syntax allows user to be more precise as to what is being
undefined:
```typeql
undefine
owns age from person;
@regex from first-name;
as name from person owns first-name;
```
New, more concise delete syntax:
```typeql
match $p isa person, has name $n;
delete $n of $p;
```

Implement JSON-like string unescaping (closes #106).

See [The TypeDB 3.0 Roadmap](<https://typedb.com/blog/typedb-3-roadmap>)
for more details!

+++ rust/common/mod.rs
@@ -14,13 +14,13 @@ pub use error::Error;
 
 pub type Result<T = ()> = std::result::Result<T, Error>;
 
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
 pub struct LineColumn {
-    pub line: usize,
-    pub column: usize,
+    pub line: u32,
+    pub column: u32,
 }
 
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
 pub struct Span {
     pub begin: LineColumn,
     pub end: LineColumn,
